pyproject.toml
[project]
name = "agent_openai"
version = "0.1.0"
description = "agent-openai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.1,<1.0.0",
    "pymongo>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[project.scripts]
agent_openai = "agent_openai.main:run"
run_crew = "agent_openai.main:run"
train = "agent_openai.main:train"
replay = "agent_openai.main:replay"
test = "agent_openai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
